Add "idle" states during heating and cooling.
Dev version of esphome supports an idle state, so we'll take a guess at
what the state is at a given point by comparing the set point.

Note that this commit does not take into account hysteresis around the
set point, and just assumes that if the temperature is at or over the
set point during heating that the unit is idle, or if the temperature is
at or under the setpoint during cooling that it's idle.